Create database tables
Follows the hand-written entity relationship diagram for Phase 1 which
can be found in docs/hand_drawn_erd.png.

Steps:
- `bin/rails generate model written_language name:text:uniq`
- `bin/rails generate model written_term name:text written_language:belongs_to`
- `bin/rails generate model category name:text:uniq`
- `bin/rails generate model video_url url:text:uniq approved:boolean written_term:belongs_to`
- `bin/rails generate model url_category up_votes:integer down_votes:integer category:belongs_to video_url:belongs_to`
Then updates the model files with the has_many relationships
- `bin/rails db:migrate`
Lots of command-line experimentation for erd-rails in order to create an
automated entity relationship diagram to compare with the hand-written
one to verify everything was made to plan. I've stored the command in
`lib/tasks/erd.sh` so it can be easily used later.
- `chmod u+x lib/tasks/erd.sh`
- `lib/tasks/erd.sh`