Default videos to not approved
The plan is that videos won't show to other users on the site until
they have been approved. This is in order to protect users against
inappropriate content. Approval is an initial check that the video is a
video of a sign, whereas votes will then be on the quality of that sign.